Alright, enough is enough.  Tired of waiting for MS to fix the WebView RE Issue #13.
We're going balls to the wall and going to parse this whole thing ourselves.  RichTextBlock here we come!

Much more work left to be done but a lot of tags work and hyperlinks launch to an external browser.

How embedding is going to work?  Well, that's a good question that I don't have an answer to.  It may have to change.